requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-dotenv==1.0.1
soupsieve==2.6
typing_extensions==4.12.2
yarl==1.15.2
